# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ApiResponseListMonzoConnectionResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/MonzoConnectionResponse"
            },
            "type": "array"
          },
          "success": {
            "type": "boolean"
          },
          "timestamp": {
            "format": "date-time",
            "type": "string"
          }
        },
        "type": "object"
      },
      "ApiResponseMonzoConnectInitResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/MonzoConnectInitResponse"
          },
          "success": {
            "type": "boolean"
          },
          "timestamp": {
            "format": "date-time",
            "type": "string"
          }
        },
        "type": "object"
      },
      "ApiResponseMonzoConnectionResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/MonzoConnectionResponse"
          },
          "success": {
            "type": "boolean"
          },
          "timestamp": {
            "format": "date-time",
            "type": "string"
          }
        },
        "type": "object"
      },
      "ApiResponseMonzoStatusResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/MonzoStatusResponse"
          },
          "success": {
            "type": "boolean"
          },
          "timestamp": {
            "format": "date-time",
            "type": "string"
          }
        },
        "type": "object"
      },
      "ApiResponseMonzoSyncProgressResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/MonzoSyncProgressResponse"
          },
          "success": {
            "type": "boolean"
          },
          "timestamp": {
            "format": "date-time",
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ApplicationContext": {
        "properties": {
          "applicationName": {
            "type": "string"
          },
          "autowireCapableBeanFactory": {
            "$ref": "#/components/schemas/AutowireCapableBeanFactory"
          },
          "beanDefinitionCount": {
            "format": "int32",
            "type": "integer"
          },
          "beanDefinitionNames": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "classLoader": {
            "properties": {
              "defaultAssertionStatus": {
                "type": "boolean",
                "writeOnly": true
              },
              "definedPackages": {
                "items": {
                  "properties": {
                    "annotations": {
                      "items": {
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "declaredAnnotations": {
                      "items": {
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "implementationTitle": {
                      "type": "string"
                    },
                    "implementationVendor": {
                      "type": "string"
                    },
                    "implementationVersion": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "sealed": {
                      "type": "boolean"
                    },
                    "specificationTitle": {
                      "type": "string"
                    },
                    "specificationVendor": {
                      "type": "string"
                    },
                    "specificationVersion": {
                      "type": "string"
                    }
                  },
                  "type": "object"
                },
                "type": "array"
              },
              "name": {
                "type": "string"
              },
              "registeredAsParallelCapable": {
                "type": "boolean"
              },
              "unnamedModule": {
                "properties": {
                  "annotations": {
                    "items": {
                      "type": "object"
                    },
                    "type": "array"
                  },
                  "declaredAnnotations": {
                    "items": {
                      "type": "object"
                    },
                    "type": "array"
                  },
                  "descriptor": {
                    "properties": {
                      "automatic": {
                        "type": "boolean"
                      },
                      "open": {
                        "type": "boolean"
                      }
                    },
                    "type": "object"
                  },
                  "layer": {
                    "type": "object"
                  },
                  "name": {
                    "type": "string"
                  },
                  "named": {
                    "type": "boolean"
                  },
                  "nativeAccessEnabled": {
                    "type": "boolean"
                  },
                  "packages": {
                    "items": {
                      "type": "string"
                    },
                    "type": "array",
                    "uniqueItems": true
                  }
                },
                "type": "object"
              }
            },
            "type": "object"
          },
          "displayName": {
            "type": "string"
          },
          "environment": {
            "$ref": "#/components/schemas/Environment"
          },
          "id": {
            "type": "string"
          },
          "parent": {
            "$ref": "#/components/schemas/ApplicationContext"
          },
          "parentBeanFactory": {
            "$ref": "#/components/schemas/BeanFactory"
          },
          "startupDate": {
            "format": "int64",
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "AutowireCapableBeanFactory": {
        "type": "object"
      },
      "BeanFactory": {
        "type": "object"
      },
...
      "DefaultHttpStatusCode": {
        "allOf": [
          {
            "$ref": "#/components/schemas/HttpStatusCode"
          }
        ]
      },
...
      "Environment": {
        "properties": {
          "activeProfiles": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "defaultProfiles": {
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
...
      "MonzoConnectInitResponse": {
        "properties": {
          "authorizationUrl": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "MonzoConnectionResponse": {
        "properties": {
          "connectedAt": {
            "format": "date-time",
            "type": "string"
          },
          "expiresAt": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          },
          "isTokenExpired": {
            "type": "boolean"
          },
          "monzoUserId": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "MonzoStatusResponse": {
        "properties": {
          "backfillStatus": {
            "enum": [
              "NOT_STARTED",
              "IN_PROGRESS",
              "NEEDS_REAUTH",
              "COMPLETED"
            ],
            "type": "string"
          },
          "connected": {
            "type": "boolean"
          },
          "connectionCount": {
            "format": "int64",
            "type": "integer"
          },
          "tokenStatus": {
            "enum": [
              "ACTIVE",
              "EXPIRING_SOON",
              "RECONNECT_REQUIRED"
            ],
            "type": "string"
          }
        },
        "type": "object"
      },
      "MonzoSyncProgressResponse": {
        "properties": {
          "accounts": {
            "items": {
              "$ref": "#/components/schemas/AccountProgress"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
...
      "RedirectView": {
        "properties": {
          "applicationContext": {
            "$ref": "#/components/schemas/ApplicationContext"
          },
          "attributes": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "writeOnly": true
          },
          "attributesCSV": {
            "type": "string",
            "writeOnly": true
          },
          "attributesMap": {
            "additionalProperties": {
              "type": "object"
            },
            "type": "object"
          },
          "beanName": {
            "type": "string"
          },
          "contentType": {
            "type": "string"
          },
          "contextRelative": {
            "type": "boolean",
            "writeOnly": true
          },
          "encodingScheme": {
            "type": "string",
            "writeOnly": true
          },
          "expandUriTemplateVariables": {
            "type": "boolean",
            "writeOnly": true
          },
          "exposeContextBeansAsAttributes": {
            "type": "boolean",
            "writeOnly": true
          },
          "exposeModelAttributes": {
            "type": "boolean",
            "writeOnly": true
          },
          "exposePathVariables": {
            "type": "boolean"
          },
          "exposedContextBeanNames": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "writeOnly": true
          },
          "hosts": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "http10Compatible": {
            "type": "boolean",
            "writeOnly": true
          },
          "propagateQueryParams": {
            "type": "boolean",
            "writeOnly": true
          },
          "propagateQueryProperties": {
            "type": "boolean"
          },
          "redirectView": {
            "type": "boolean"
          },
          "requestContextAttribute": {
            "type": "string"
          },
          "servletContext": {
            "$ref": "#/components/schemas/ServletContext",
            "writeOnly": true
          },
          "staticAttributes": {
            "additionalProperties": {
              "type": "object"
            },
            "type": "object"
          },
          "statusCode": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/DefaultHttpStatusCode"
              },
              {
                "$ref": "#/components/schemas/HttpStatus"
              }
            ],
            "writeOnly": true
          },
          "url": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "User": {
        "properties": {
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "email": {
            "maxLength": 255,
            "minLength": 0,
            "type": "string"
          },
          "emailVerified": {
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Personal Monzo-integrated budgeting app. All responses use the ApiResponse/ApiError envelope; authentication is a JWE session cookie obtained via the magic-link flow (dev: POST /api/dev/auth/quick-login). Unauthenticated requests to protected routes return 401 with an ApiError envelope (code MISSING_TOKEN).",
    "title": "Budgeteer API",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/monzo/callback": {
      "get": {
        "operationId": "handleCallback",
        "parameters": [
          {
            "in": "query",
            "name": "code",
            "required": false,
            "schema": {
              "maxLength": 1024,
              "minLength": 0,
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "state",
            "required": true,
            "schema": {
              "maxLength": 64,
              "minLength": 0,
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "error",
            "required": false,
            "schema": {
              "maxLength": 256,
              "minLength": 0,
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "error_description",
            "required": false,
            "schema": {
              "maxLength": 1000,
              "minLength": 0,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseMonzoConnectionResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "monzo-controller"
        ]
      }
    },
    "/api/v1/monzo/connect": {
      "get": {
        "operationId": "initiateOAuth",
        "parameters": [
          {
            "in": "query",
            "name": "user",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/User"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RedirectView"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "monzo-controller"
        ]
      },
      "post": {
        "operationId": "initiateOAuthJson",
        "parameters": [
          {
            "in": "query",
            "name": "user",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/User"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseMonzoConnectInitResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "monzo-controller"
        ]
      }
    },
    "/api/v1/monzo/connections": {
      "get": {
        "operationId": "listConnections",
        "parameters": [
          {
            "in": "query",
            "name": "user",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/User"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseListMonzoConnectionResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "monzo-controller"
        ]
      }
    },
    "/api/v1/monzo/connections/{id}": {
      "delete": {
        "operationId": "disconnectConnection",
        "parameters": [
          {
            "in": "query",
            "name": "user",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/User"
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "monzo-controller"
        ]
      },
      "get": {
        "operationId": "getConnection",
        "parameters": [
          {
            "in": "query",
            "name": "user",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/User"
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseMonzoConnectionResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "monzo-controller"
        ]
      }
    },
    "/api/v1/monzo/status": {
      "get": {
        "operationId": "getStatus",
        "parameters": [
          {
            "in": "query",
            "name": "user",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/User"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseMonzoStatusResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "monzo-controller"
        ]
      }
    },
    "/api/v1/monzo/sync": {
      "post": {
        "operationId": "syncNow",
        "parameters": [
          {
            "in": "query",
            "name": "user",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/User"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseMonzoSyncProgressResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "monzo-controller"
        ]
      }
    },
    "/api/v1/monzo/sync/progress": {
      "get": {
        "operationId": "getSyncProgress",
        "parameters": [
          {
            "in": "query",
            "name": "user",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/User"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseMonzoSyncProgressResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "monzo-controller"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Generated server url",
      "url": "http://localhost:8080"
    }
  ]
}